project: rename paths from build to src
