+++ test/functional/boleto/create/bradesco/refused.js
@@ -3,8 +3,8 @@ import Promise from 'bluebird'
 import { assert } from '../../../../helpers/chai'
 import { normalizeHandler } from '../../../../helpers/normalizer'
 import { mock, mockFunction, restoreFunction } from '../../../../helpers/boleto'
-import * as boletoHandler from '../../../../../build/resources/boleto'
-import * as Provider from '../../../../../build/providers/bradesco'
+import * as boletoHandler from '../../../../../src/resources/boleto'
+import * as Provider from '../../../../../src/providers/bradesco'
 
 const create = normalizeHandler(boletoHandler.create)
