+++ test/functional/boleto/create/development/pending.js
@@ -3,7 +3,7 @@ import Promise from 'bluebird'
 import { assert } from '../../../../helpers/chai'
 import { normalizeHandler } from '../../../../helpers/normalizer'
 import { mock } from '../../../../helpers/boleto'
-import * as boletoHandler from '../../../../../build/resources/boleto'
+import * as boletoHandler from '../../../../../src/resources/boleto'
 
 const create = normalizeHandler(boletoHandler.create)
