+++ test/functional/boleto/index.js
@@ -2,8 +2,8 @@ import test from 'ava'
 import { assert } from '../../helpers/chai'
 import { normalizeHandler } from '../../helpers/normalizer'
 import { createBoleto, userQueueUrl } from '../../helpers/boleto'
-import * as boletoHandler from '../../../build/resources/boleto'
-import { getModel } from '../../../build/database'
+import * as boletoHandler from '../../../src/resources/boleto'
+import { getModel } from '../../../src/database'
 
 let Boleto
