+++ test/unit/lib/http/response.js
@@ -3,7 +3,7 @@ import {
   buildSuccessResponse,
   buildFailureResponse,
   buildErrorPayload
-} from '../../../../build/lib/http/response'
+} from '../../../../src/lib/http/response'
 
 test('buildSuccessResponse', async (t) => {
   const input = { message: 'This is some useful message' }
